feat(router): using recommended file-based router
when this repo was created @tanstack/router was still experimental and using code-based routing.
File-based routing was barely in its infancy. After 6 or so months, @tanstack/router was updated to
version 1.0+ and renamed to @tanstack/react-router. This updated version now recommends the use of
File-based routing due to the simplicity provided with similar and/or better results in less code.

BREAKING CHANGE: code-based routing replaced with file-based routing